+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to mmap-rs will be documented in this file.
 - Document flags that may have no operation on platforms that do not support the flag, such as `MmapFlags::STACK` on Microsoft Windows.
 - Use `mach_vm_region_recurse()` on MacOS to recurse into submappings.
 - Change `MmapFlags::COPY_ON_WRITE` to `MmapFlags::SHARED`, such that private memory mappings are the default and since copy-on-write semantics only make sense for file mappings.
+- Fix issue where `MemoryAreas::query()` and related functions returned the wrong `ShareMode`.
 
 ## 0.5.0